Fix drop order in `IntoIter`
Technically, the old code violates the soundness requirements of
`ExclusiveRef`. However, because `ExclusiveRef` is kept private to the
crate, we can show that this issue cannot cause any undefined behavior:
although the code that drops `_tree` (`BTreeVec::drop`) creates an
`ExclusiveRef` while `leaf` still exists, it does not access `leaf` in
any way. Thus, the crate is still sound from the perspective of an
external user, and previous releases do not need to be yanked.